To fix this warning, use size_t instead of type. Alternatively, use an integral type that is at least as large as size_t.

+++ GP/Common.cpp
@@ -164,8 +164,8 @@ std::vector<std::string> CCommon::split(std::string str, std::string pattern)
     std::string::size_type pos;
     std::vector<std::string> result;
     str += pattern;//扩展字符串以方便操作
-    int size = str.size();
-    for (int i = 0; i < size; i++)
+    size_t size = str.size();
+    for (size_t i = 0; i < size; i++)
     {
         pos = str.find(pattern, i);
         if (pos < size)
